+++ NEWS.md
@@ -1,3 +1,15 @@
+# AMAPVox 1.0.0
+
+*2023-03-15*
+From binary AMAPVox 2.0, no need to install local Java VM, it is embedded within AMAPVox.
+Released on CRAN
+
+# AMAPVox 0.13.0
+
+*2023-01-26*
+Get AMAPVox binaries from forge.ird.fr with Gitlab API
+Not released on CRAN.
+
 # AMAPVox 0.12.1
 
 *2022-12-05*